fixed windows bug

+++ takenote/safefile.py
@@ -64,7 +64,8 @@ class SafeFile (file):
 
         if self._tmp:
             # NOTE: windows will not allow rename when destination file exists
-            os.remove(self._filename)
+            if os.path.exists(self._filename):
+                os.remove(self._filename)
             os.rename(self._tmp, self._filename)
             self._tmp = None
